moved prevector
